Create RanDom Number Generator
It's a Lock System with keywords. A user does not have to remember the keywords. Instead of that he/she has to remember the index number of his lock Matrix. Different number will generate each time in same index and the user have to input them in a box to get through.